% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bmi3.R
\name{bmi3}
\alias{bmi3}
\title{This function creates bmi groups}
\usage{
bmi3(x)
}
\arguments{
\item{x}{numeric vector}
}
\value{
Factor Variable
}
\description{
This function creates bmi groups
}
\examples{
bmi.vals <- rnorm (n=20, mean=25, sd=3 )
bmi3(bmi.vals)

 @export
}
