% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mybin.df.R
\name{mybin}
\alias{mybin}
\title{Binomial simulation}
\usage{
mybin(iter = 100, n = 10, p = 0.7)
}
\arguments{
\item{iter}{: number of iterartions}

\item{n}{: number of trials}

\item{p}{: probability of success}
}
\value{
: a plot of the simmulations with proportions & a table of proportions
}
\description{
This uses the sample function to produce a binomial simmulation
}
\examples{
mybin()

}
