% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Package2020.R
\docType{Package}
\name{Package2020}
\alias{Package2020}
\title{BMI package}
\description{
Contains a function that does does something fantastic
}
\author{
Wagner Rosa (following exampple by Dane Van Domelen on YT ) \email{wagner.e.rosa-1@ou.edu}
}
